消除警告

+++ Mod/Utils/Utils.h
@@ -511,12 +511,12 @@ public:
 			return std::string(ansi);
 		}
 		int len = MultiByteToWideChar(CP_ACP, 0, ansi, -1, NULL, 0);
-		wchar_t* wstr = new wchar_t[len + 1];
-		memset(wstr, 0, static_cast<size_t>(len) + 1);
+		wchar_t* wstr = new wchar_t[(size_t)len + 1];
+		memset(wstr, 0, (size_t)len + 1);
 		MultiByteToWideChar(CP_ACP, 0, ansi, -1, wstr, len);
 		len = WideCharToMultiByte(CP_UTF8, 0, wstr, -1, NULL, 0, NULL, NULL);
-		char* str = new char[len + 1];
-		memset(str, 0, static_cast<size_t>(len) + 1);
+		char* str = new char[(size_t)len + 1];
+		memset(str, 0, (size_t)len + 1);
 		WideCharToMultiByte(CP_UTF8, 0, wstr, -1, str, len, NULL, NULL);
 		if (wstr) delete[] wstr;
 		std::string ret = str;
@@ -530,18 +530,16 @@ public:
 			return std::string(utf8);
 		}
 		int len = MultiByteToWideChar(CP_UTF8, 0, utf8, -1, NULL, 0);
-		wchar_t* wstr = new wchar_t[len + 1];
-		memset(wstr, 0, static_cast<size_t>(len) + 1);
+		wchar_t* wstr = new wchar_t[(size_t)len + 1];
+		memset(wstr, 0, (size_t)len + 1);
 		MultiByteToWideChar(CP_UTF8, 0, utf8, -1, wstr, len);
 		len = WideCharToMultiByte(CP_ACP, 0, wstr, -1, NULL, 0, NULL, NULL);
-		char* str = new char[len + 1];
-		memset(str, 0, static_cast<size_t>(len) + 1);
+		char* str = new char[(size_t)len + 1];
+		memset(str, 0, (size_t)len + 1);
 		WideCharToMultiByte(CP_ACP, 0, wstr, -1, str, len, NULL, NULL);
 		if (wstr) delete[] wstr;
 		std::string ret = str;
 		if (str) delete[] str;
 		return ret;
 	}
-
-
 };
